+++ priv/templates/curator.install/curator_hooks.ex
@@ -1,4 +1,0 @@
-defmodule <%= base %>.CuratorHooks do
-  use <%= base %>.Web, :controller
-  use Curator.Hooks
-end