+++ test/curator/session_handler_test.exs
@@ -1,4 +1,0 @@
-defmodule Curator.SessionHandlerTest do
-  use ExUnit.Case, async: true
-  doctest Curator.SessionHandler
-end