+++ test/curator/view_helpers_test.exs
@@ -1,4 +1,0 @@
-defmodule Curator.ViewHelpersTest do
-  use ExUnit.Case, async: true
-  doctest Curator.ViewHelpers
-end