+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v 0.2.0
+Rewrite - currently only supports an Ueberauth workflow (and timeouts). v0.3.0 will add support back for database authentication.
+
+* Generators match Phoenix 1.3 structure
+* Requires Guardian 1.0
+* Configuration is now done at the module level (following the pattern established in Guardian)
+* Ueberauth module & generator
+* Timeoutable module & generator
+
 # v 0.1.0
 
 Initial Release